+++ libs/core/pipes/shorten-name/src/index.ts
@@ -1,0 +1,3 @@
+export * from './lib/shorten-name.module';
+
+export * from './lib/shorten-name.module';