+++ libs/core/pipes/shorten-name/src/lib/shorten-name.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { ShortenNamePipe } from './shorten-name.pipe';
+
+@NgModule({
+  imports: [CommonModule],
+  declarations: [ShortenNamePipe],
+  exports: [ShortenNamePipe],
+})
+export class ShortenNamePipeModule {}