+++ libs/web/calendar/ui/calendar-header/src/lib/calendar-header.component.ts
@@ -4,17 +4,15 @@ import {
   Component,
   Inject,
   Input,
+  ViewChild,
 } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { ScheduleComponent, View } from '@syncfusion/ej2-angular-schedule';
 import { TuiDestroyService, TuiMonth } from '@taiga-ui/cdk';
 import {
-  TuiAppearance,
+  tuiButtonOptionsProvider,
   TuiNotificationsService,
-  TUI_BUTTON_OPTIONS,
 } from '@taiga-ui/core';
-import { PermissionConstant } from '@teaching-scheduling-system/core/data-access/constants';
-import { Nullable } from '@teaching-scheduling-system/core/data-access/models';
 import { fadeIn } from '@teaching-scheduling-system/core/ui/animations';
 import {
   DateHelper,
@@ -22,28 +20,20 @@ import {
 } from '@teaching-scheduling-system/core/utils/helpers';
 import {
   calendarChangeMonth,
-  calendarChangeSelectingState,
   calendarChangeView,
-  calendarFilter,
   calendarNext,
   calendarPrev,
   calendarSelectFilter,
-  calendarSelectModules,
   calendarSelectMonth,
   calendarSelectSelectedDate,
-  calendarSelectTeachers,
   calendarSelectView,
   CalendarState,
 } from '@teaching-scheduling-system/web/calendar/data-access';
-import {
-  CalendarFilter,
-  SimpleModel,
-  Teacher,
-} from '@teaching-scheduling-system/web/shared/data-access/models';
+import { CalendarFilterComponent } from '@teaching-scheduling-system/web/calendar/ui/calendar-filter';
+import { CalendarFilter } from '@teaching-scheduling-system/web/shared/data-access/models';
 import {
   AppShellState,
   selectNameTitle,
-  selectTeacher,
 } from '@teaching-scheduling-system/web/shared/data-access/store';
 import {
   combineLatest,
@@ -64,19 +54,19 @@ import {
   animations: [fadeIn],
   providers: [
     TuiDestroyService,
-    {
-      provide: TUI_BUTTON_OPTIONS,
-      useValue: {
-        shape: null,
-        appearance: TuiAppearance.Flat,
-        size: 'xs',
-      },
-    },
+    tuiButtonOptionsProvider({
+      appearance: 'flat',
+      size: 'xs',
+    }),
   ],
 })
 export class CalendarHeaderComponent implements AfterViewInit {
   /** INPUT */
   @Input() public scheduleComponent!: ScheduleComponent;
+
+  /** VIEWCHILD */
+  @ViewChild(CalendarFilterComponent, { static: false })
+  public filter!: CalendarFilterComponent;
 
   /** PUBLIC PROPERTIES */
   public view$: Observable<View>;
@@ -86,21 +76,12 @@ export class CalendarHeaderComponent implements AfterViewInit {
   public dateRange$!: Observable<string>;
   public openFilter = false;
   public activeToday$!: Observable<boolean>;
-  public teachers$: Observable<SimpleModel[]>;
-  public modules$: Observable<string[]>;
-  public user$: Observable<Nullable<Teacher>>;
 
   public readonly clickToday$ = new Subject<void>();
-  public readonly permissionConstant = PermissionConstant;
-
-  public showDepartmentSchedule = false;
-  public filteredTeachers: SimpleModel[] = [];
-  public filteredModules: string[] = [];
 
   /** PRIVATE PROPERTIES */
   private selectedDate$: Observable<Date>;
   private canDisplayNotification = true;
-  private readonly closeFilter$ = new Subject<void>();
   private readonly displayNotification$ = new Subject<void>();
 
   /** CONSTRUCTOR */
@@ -111,9 +92,6 @@ export class CalendarHeaderComponent implements AfterViewInit {
     private readonly appShellStore: Store<AppShellState>,
     private readonly destroy$: TuiDestroyService
   ) {
-    this.user$ = appShellStore
-      .select(selectTeacher)
-      .pipe(takeUntil(this.destroy$));
     this.selectedDate$ = store
       .select(calendarSelectSelectedDate)
       .pipe(takeUntil(this.destroy$));
@@ -126,19 +104,12 @@ export class CalendarHeaderComponent implements AfterViewInit {
     this.view$ = store
       .select(calendarSelectView)
       .pipe(takeUntil(this.destroy$));
-    this.teachers$ = store
-      .select(calendarSelectTeachers)
-      .pipe(takeUntil(this.destroy$));
-    this.modules$ = store
-      .select(calendarSelectModules)
-      .pipe(takeUntil(this.destroy$));
 
-    this.triggerFilter();
     this.handleClickToday();
     this.handleDisplayNotification();
   }
 
-  /** LIFE CYCLE */
+  /** LIFECYCLE */
   public ngAfterViewInit(): void {
     this.triggerDateRange();
     this.triggerActiveToday();
@@ -168,35 +139,8 @@ export class CalendarHeaderComponent implements AfterViewInit {
 
   public onFilterOpenChange(open: boolean): void {
     if (!open) {
-      this.closeFilter$.next();
+      this.filter?.reset();
     }
-  }
-
-  public onSelectingDepartmentChange(showDepartmentSchedule: boolean): void {
-    this.store.dispatch(
-      calendarChangeSelectingState({ changes: { showDepartmentSchedule } })
-    );
-    this.filteredModules = [];
-  }
-
-  public onSelectingTeachersChange(teachers: SimpleModel[]): void {
-    this.store.dispatch(
-      calendarChangeSelectingState({ changes: { teachers } })
-    );
-    this.filteredModules = [];
-  }
-
-  public filter(): void {
-    this.store.dispatch(
-      calendarFilter({
-        filter: {
-          showDepartmentSchedule: this.showDepartmentSchedule,
-          teachers: this.filteredTeachers,
-          modules: this.filteredModules,
-        },
-      })
-    );
-    this.openFilter = false;
   }
 
   public onClickMonth(): void {
@@ -211,13 +155,11 @@ export class CalendarHeaderComponent implements AfterViewInit {
     this.store.dispatch(calendarChangeView({ view: 'Day' }));
   }
 
-  /** PRIVATE METHODS */
-  private triggerFilter(): void {
-    combineLatest([this.filter$, this.closeFilter$])
-      .pipe(tap(([filter]) => this.resetFilter(filter)))
-      .subscribe();
+  public onFilter(): void {
+    this.openFilter = false;
   }
 
+  /** PRIVATE METHODS */
   private triggerDateRange(): void {
     this.dateRange$ = combineLatest([this.view$, this.selectedDate$]).pipe(
       map(([view]) => view),
@@ -241,8 +183,7 @@ export class CalendarHeaderComponent implements AfterViewInit {
     this.clickToday$
       .pipe(
         withLatestFrom(this.view$, this.filter$),
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        tap(([_, view, filter]) => {
+        tap(({ 1: view, 2: filter }) => {
           const today = new Date();
 
           if (
@@ -272,8 +213,7 @@ export class CalendarHeaderComponent implements AfterViewInit {
     this.displayNotification$
       .pipe(
         withLatestFrom(this.appShellStore.select(selectNameTitle)),
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        tap(([_, nameTitle]) => {
+        tap(({ 1: nameTitle }) => {
           this.canDisplayNotification = false;
           const schedule = this.scheduleComponent;
           const now = new Date();
@@ -306,17 +246,14 @@ export class CalendarHeaderComponent implements AfterViewInit {
   private triggerActiveToday(): void {
     this.activeToday$ = combineLatest([this.selectedDate$, this.view$]).pipe(
       delay(0),
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
-      map(([_, view]) =>
+      map(({ 1: view }) =>
         ScheduleHelper.dayInCurrentView(this.scheduleComponent, view)
       )
     );
   }
 
-  private resetFilter(filter: CalendarFilter): void {
-    this.showDepartmentSchedule = filter.showDepartmentSchedule;
-    this.filteredTeachers = filter.teachers;
-    this.filteredModules = filter.modules;
+  private resetFilter(): void {
+    // this.filter.close();
   }
 
   private monthDateRange(): string {