+++ libs/web/calendar/ui/calendar-menu/src/index.ts
@@ -1,0 +1,3 @@
+export * from './lib/calendar-menu.module';
+
+export * from './lib/calendar-menu.component';