+++ libs/web/exam/ui/assign-edit-dialog/src/lib/assign-edit-dialog.component.ts
@@ -9,10 +9,10 @@ import { Store } from '@ngrx/store';
 import { TuiDestroyService } from '@taiga-ui/cdk';
 import {
   TuiAppearance,
+  tuiButtonOptionsProvider,
   TuiDialogContext,
   TuiNotification,
   TuiNotificationsService,
-  TUI_BUTTON_OPTIONS,
   TUI_TEXTFIELD_APPEARANCE,
 } from '@taiga-ui/core';
 import { EApiStatus } from '@teaching-scheduling-system/web/shared/data-access/enums';
@@ -35,14 +35,10 @@ import { AssignEditExamDialogStore } from './store';
   providers: [
     AssignEditExamDialogStore,
     TuiDestroyService,
-    {
-      provide: TUI_BUTTON_OPTIONS,
-      useValue: {
-        shape: null,
-        appearance: TuiAppearance.Primary,
-        size: 'm',
-      },
-    },
+    tuiButtonOptionsProvider({
+      appearance: 'primary',
+      size: 'm',
+    }),
     {
       provide: TUI_TEXTFIELD_APPEARANCE,
       useValue: TuiAppearance.Textfield,