+++ libs/web/shell/ui/navbar/src/lib/navbar.component.ts
@@ -1,8 +1,10 @@
 import {
   ChangeDetectionStrategy,
+  ChangeDetectorRef,
   Component,
   HostListener,
   Inject,
+  TemplateRef,
 } from '@angular/core';
 import { Router } from '@angular/router';
 import { Store } from '@ngrx/store';
@@ -19,7 +21,8 @@ import {
   selectTeacher,
 } from '@teaching-scheduling-system/web/shared/data-access/store';
 import { Observable, takeUntil } from 'rxjs';
-import { NAVBAR_OPTIONS, NavbarOptions } from './navbar.token';
+import { NavbarService } from './navbar.service';
+import { NavbarOptions, NAVBAR_OPTIONS } from './navbar.token';
 
 @Component({
   selector: 'tss-navbar',
@@ -31,18 +34,22 @@ import { NAVBAR_OPTIONS, NavbarOptions } from './navbar.token';
 export class NavbarComponent {
   /** PUBLIC PROPERTIES */
   public readonly items = NavbarConstants.items;
-  public user$: Observable<Nullable<Teacher>> | undefined;
+  public readonly rightMenu$: Observable<Nullable<TemplateRef<never>>>;
+
   public openMobileNav = false;
   public openDropDown = false;
   public isMobileScreen = true;
+  public user$: Observable<Nullable<Teacher>> | undefined;
 
   /** CONSTRUCTOR */
   constructor(
     private readonly router: Router,
     private readonly accessTokenService: AccessTokenService,
     private readonly authService: AuthService,
+    private readonly cdr: ChangeDetectorRef,
     @Inject(NAVBAR_OPTIONS) public readonly options: NavbarOptions,
     appShellStore: Store<AppShellState>,
+    navbarService: NavbarService,
     destroy$: TuiDestroyService
   ) {
     this.onResize();
@@ -51,6 +58,7 @@ export class NavbarComponent {
         .select(selectTeacher)
         .pipe(takeUntil(destroy$));
     }
+    this.rightMenu$ = navbarService.rightMenu$.pipe(takeUntil(destroy$));
   }
 
   /** PUBLIC METHODS */
@@ -74,7 +82,10 @@ export class NavbarComponent {
     }
   }
 
-  public toggleMobileNav(open: boolean): void {
+  public toggleMobileNav(open: boolean, needMarkForCheck = false): void {
     this.openMobileNav = open;
+    if (needMarkForCheck) {
+      this.cdr.markForCheck();
+    }
   }
 }