+++ libs/web/statistic/ui/change-schedule-filter/src/lib/change-schedule-filter.component.ts
@@ -92,7 +92,7 @@ export class ChangeScheduleFilterComponent implements OnInit {
     this.handleExport();
   }
 
-  /** LIFE CYCLE */
+  /** LIFECYCLE */
   public ngOnInit(): void {
     this.statisticizeFirstTime();
   }